+++ CHANGELOG.md
@@ -99,6 +99,7 @@ with the `YY.0M.MICRO` scheme.
 - Add `pico.raft_term()`
 - Add `pico.change_password()`
 - Add `pico.wait_ddl_finalize()`
+- Make `pico.cas` follow access control rules
 - Change `pico.raft_log()` arguments
 - Make `opts.timeout` optional in most functions